% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimoT.R
\docType{package}
\name{nimoT-package}
\alias{nimoT}
\alias{nimoT-package}
\title{nimoT: A package for reading and visualising nimoT data}
\description{
This is a shiny project
}
\author{
\strong{Maintainer}: Emanuel Huber \email{emanuel.huber@pm.me}

}
